+++ src/IQuery.ts
@@ -1,5 +1,5 @@
 import type { TaskLayoutOptions } from './Layout/TaskLayoutOptions';
-import type { QueryLayoutOptions } from './QueryLayoutOptions';
+import type { QueryLayoutOptions } from './Layout/QueryLayoutOptions';
 import type { Task } from './Task/Task';
 import type { Grouper } from './Query/Group/Grouper';
 import type { QueryResult } from './Query/QueryResult';
