+++ src/TaskSerializer/DefaultTaskSerializer.ts
@@ -1,6 +1,6 @@
 import type { Moment } from 'moment';
 import { TaskLayoutOptions } from '../Layout/TaskLayoutOptions';
-import type { TaskLayoutComponent } from '../TaskLayout';
+import type { TaskLayoutComponent } from '../Layout/TaskLayout';
 import { Recurrence } from '../Task/Recurrence';
 import { Priority, Task, TaskRegularExpressions } from '../Task/Task';
 import type { TaskDetails, TaskSerializer } from '.';
