+++ tests/TaskLayout.test.ts
@@ -3,8 +3,8 @@
  */
 
 import { TaskLayoutOptions } from '../src/Layout/TaskLayoutOptions';
-import { QueryLayoutOptions } from '../src/QueryLayoutOptions';
-import { TaskLayout } from '../src/TaskLayout';
+import { QueryLayoutOptions } from '../src/Layout/QueryLayoutOptions';
+import { TaskLayout } from '../src/Layout/TaskLayout';
 
 describe('TaskLayout tests', () => {
     it('should generate expected CSS classes for default layout', () => {
