+++ tests/TaskLineRenderer.test.ts
@@ -7,10 +7,10 @@ import { GlobalFilter } from '../src/Config/GlobalFilter';
 import { resetSettings, updateSettings } from '../src/Config/Settings';
 import { TaskLayoutOptions } from '../src/Layout/TaskLayoutOptions';
 import { DateParser } from '../src/Query/DateParser';
-import { QueryLayoutOptions } from '../src/QueryLayoutOptions';
+import { QueryLayoutOptions } from '../src/Layout/QueryLayoutOptions';
 import type { Task } from '../src/Task/Task';
 import { TaskRegularExpressions } from '../src/Task/Task';
-import { type TaskLayoutComponent, taskLayoutComponents } from '../src/TaskLayout';
+import { type TaskLayoutComponent, taskLayoutComponents } from '../src/Layout/TaskLayout';
 import type { TextRenderer } from '../src/TaskLineRenderer';
 import { TaskLineRenderer } from '../src/TaskLineRenderer';
 import { fromLine } from './TestingTools/TestHelpers';
